Split scores export wire shaping and classify-then-materialize precedence.
Extract export_wire.py to break the precedence/materialization cycle, replace the dual-mode resolver with classify_scores_export plus build_scores_export_payload, and make is_persisted branch-aware for authoritative persisted rows only.
