+++ packages/api/api/analytics/scores/export_materialization.py
@@ -4,90 +4,25 @@ from __future__ import annotations
 
 from api.analytics.export_context import AnalyticQueryContext
 from api.analytics.export_types import ExportScope
-from api.analytics.military_score_inference.actions import ActionCatalog
-from api.analytics.military_score_inference.inference_api_payload import (
-    serialize_solution_without_arithmetic,
-    serialize_solutions_with_arithmetic,
-)
 from api.analytics.military_score_inference.inference_stream_rows import (
-    CachedCompleteRowAdmission,
-    ImmediateRowAdmission,
-    RowStreamAdmission,
     resolve_row_stream_admission,
 )
 from api.analytics.military_score_inference.inference_stream_scope import InferenceStreamScope
-from api.analytics.military_score_inference.models import InferenceObservation, InferenceSolution
-from api.analytics.military_score_inference.row_run import RowRun
-
-# Re-export precedence APIs for stable import paths (exports.py, tests).
-from api.analytics.scores.export_precedence import (  # noqa: F401
+from api.analytics.scores.export_precedence import (
     PERSISTABLE_INFERENCE_STATUSES,
-    ScoresExportPayload,
     ScoresInferenceSnapshot,
     SearchStatus,
-    is_scores_inference_ensure_satisfied,
-    resolve_scores_export_payload,
-    resolve_scores_export_search_status,
-    scores_export_precedence_branch,
 )
 from api.analytics.scores.export_services import ScoresExportContext
-from api.serialization.inference_row_persistence import PersistedInferenceRow
 
 __all__ = [
-    "ScoresExportPayload",
     "ScoresInferenceSnapshot",
     "export_meta_branch",
     "gather_scores_inference_snapshot",
     "hull_catalog_mask_branch",
     "is_persistable_inference_status",
-    "is_scores_inference_ensure_satisfied",
-    "ranked_solutions_from_wire",
-    "resolve_scores_export_payload",
-    "resolve_scores_export_search_status",
-    "scores_export_precedence_branch",
     "scores_inference_stream_scope",
-    "solutions_diagnostics_from_wire_complete_event",
-    "_solutions_from_admission_or_scheduler",
-    "solutions_from_domain",
 ]
-
-
-def ranked_solutions_from_wire(
-    wire_solutions: list[dict[str, object]],
-) -> list[dict[str, object]]:
-    """Return held solutions in rank order, matching inference row wire shape."""
-    return sorted(
-        wire_solutions,
-        key=lambda solution: int(solution.get("objectiveValue", 0)),
-        reverse=True,
-    )
-
-
-def solutions_diagnostics_from_wire_complete_event(
-    wire_event: dict[str, object],
-) -> tuple[list[dict[str, object]], dict[str, object] | None, int]:
-    """Extract solutions, diagnostics, and held count from a wire complete event."""
-    wire_solutions = wire_event.get("solutions")
-    solutions = ranked_solutions_from_wire(
-        wire_solutions if isinstance(wire_solutions, list) else []
-    )
-    event_diagnostics = wire_event.get("diagnostics")
-    diagnostics = event_diagnostics if isinstance(event_diagnostics, dict) else None
-    solutions_held = int(wire_event.get("solutionCount", 0))
-    return solutions, diagnostics, solutions_held
-
-
-def solutions_from_domain(
-    solutions: list[InferenceSolution] | tuple[InferenceSolution, ...],
-    *,
-    observation: InferenceObservation | None = None,
-    catalog: ActionCatalog | None = None,
-) -> list[dict[str, object]]:
-    """Serialize held domain solutions using the same shape as inference row wire."""
-    ranked = sorted(solutions, key=lambda solution: solution.objective_value, reverse=True)
-    if observation is not None and catalog is not None:
-        return serialize_solutions_with_arithmetic(observation, catalog, ranked)
-    return [serialize_solution_without_arithmetic(solution) for solution in ranked]
 
 
 def export_meta_branch(
@@ -109,106 +44,6 @@ def hull_catalog_mask_branch(enabled_hull_ids: frozenset[int] | set[int]) -> dic
     return {"enabledHullIds": sorted(enabled_hull_ids)}
 
 
-def terminal_row_admission(
-    admission: RowStreamAdmission | None,
-) -> ImmediateRowAdmission | CachedCompleteRowAdmission | None:
-    """Return admission only when it carries a terminal wire-complete payload."""
-    if isinstance(admission, ImmediateRowAdmission) and admission.events:
-        return admission
-    if isinstance(admission, CachedCompleteRowAdmission) and admission.event is not None:
-        return admission
-    return None
-
-
-def _diagnostics_from_scheduler_ladder(scheduler_run: RowRun) -> dict[str, object] | None:
-    """Build scores row inference diagnostics wire from live scheduler ladder state."""
-    ladder_state = scheduler_run.ladder_state
-    assert ladder_state is not None
-    if (
-        not ladder_state.last_diagnostics
-        and ladder_state.catalog is None
-        and not ladder_state.step_diagnostics
-        and not ladder_state.policy_steps_attempted
-    ):
-        return None
-
-    from api.analytics.military_score_inference.analytic import build_inference_solver_diagnostics
-
-    session = scheduler_run.session
-    solver_diagnostics: dict[str, object] = {
-        "status": ladder_state.last_status,
-        **ladder_state.last_diagnostics,
-    }
-    extra: dict[str, object] = {
-        "solution_count": len(ladder_state.merged_solutions),
-    }
-    if ladder_state.policy_steps_attempted:
-        extra["policy_steps_attempted"] = list(ladder_state.policy_steps_attempted)
-    if ladder_state.step_diagnostics:
-        extra["policy_step_attempts"] = list(ladder_state.step_diagnostics)
-
-    return build_inference_solver_diagnostics(
-        turn=session.turn.settings.turn,
-        observation=session.observation,
-        problem=ladder_state.problem,
-        catalog=ladder_state.catalog,
-        turn_info=session.turn,
-        solver=solver_diagnostics,
-        extra=extra,
-    )
-
-
-def _solutions_from_scheduler_ladder(
-    scheduler_run: RowRun,
-) -> tuple[list[dict[str, object]], dict[str, object] | None, int]:
-    ladder_state = scheduler_run.ladder_state
-    assert ladder_state is not None
-    merged = ladder_state.merged_solutions
-    return (
-        solutions_from_domain(
-            merged,
-            observation=scheduler_run.session.observation,
-            catalog=ladder_state.catalog,
-        ),
-        _diagnostics_from_scheduler_ladder(scheduler_run),
-        len(merged),
-    )
-
-
-def _solutions_from_admission_or_scheduler(
-    *,
-    admission: RowStreamAdmission | None,
-    scheduler_run: RowRun | None,
-    persisted_row: PersistedInferenceRow | None,
-) -> tuple[list[dict[str, object]], dict[str, object] | None, int]:
-    if isinstance(admission, ImmediateRowAdmission) and admission.events:
-        return solutions_diagnostics_from_wire_complete_event(admission.events[-1])
-    if isinstance(admission, CachedCompleteRowAdmission) and admission.event is not None:
-        return solutions_diagnostics_from_wire_complete_event(admission.event)
-    if scheduler_run is not None and scheduler_run.ladder_state is not None:
-        return _solutions_from_scheduler_ladder(scheduler_run)
-    return (
-        [],
-        None,
-        held_solution_count(
-            persisted_row=persisted_row,
-            scheduler_run=scheduler_run,
-        ),
-    )
-
-
-def held_solution_count(
-    *,
-    persisted_row: PersistedInferenceRow | None,
-    scheduler_run: RowRun | None,
-) -> int:
-    if persisted_row is not None:
-        return persisted_row.solution_count
-    if scheduler_run is not None and scheduler_run.ladder_state is not None:
-        return len(scheduler_run.ladder_state.merged_solutions)
-    return 0
-
-
 def is_persistable_inference_status(status: str) -> bool:
     return status in PERSISTABLE_INFERENCE_STATUSES
 
@@ -224,7 +59,7 @@ def scores_inference_stream_scope(scope: ExportScope) -> InferenceStreamScope:
 def _load_persisted_row(
     services: ScoresExportContext,
     scope: ExportScope,
-) -> PersistedInferenceRow | None:
+):
     if services.persistence is None or scope.player_id is None:
         return None
     return services.persistence.get_row(
